+++ src/components/Icon/Svg/Svg.tsx
@@ -1,15 +1,9 @@
 export interface SvgProps {
-  /** CSS styling class */
   className?: string;
-
-  /** The SVG image data */
   path: string;
-
-  /** Size of the SVG */
   viewbox: string;
 }
 
-/** Renders an SVG image */
 export const Svg = ({ className, path, viewbox }: SvgProps): JSX.Element => (
   <svg
     className={className}
